Spam_classifier
Welcome to the Spam Classifier repository powered by the Naive Bayes algorithm and the NLTK (Natural Language Toolkit) library! This project is dedicated to building an efficient classifier for identifying and filtering spam messages. Dive into the world of machine learning and NLP, exploring the robust capabilities of NLTK for text processing. Discover how to preprocess, train, and fine-tune the Naive Bayes model to deliver precise spam detection and improve email security. Join us on this journey to leverage the power of NLTK and Naive Bayes for effective spam filtering and streamlined communication.